refactor: improve type imports for login and verify OTP modules
